fix a bug when it asks to select the path and a crash when entering 999 into ping

+++ ExePatcher.cpp
@@ -742,7 +742,7 @@ bool overwriteFunction(HWND mainWindow, char* funcPtr, char* textSectionEnd, int
         return false;
 	}
 	bool nothingWasChanged = false;
-	int positionToWriteFrom = funcPtr - wholeFileBegin;
+	int positionToWriteFrom = (int)(funcPtr - wholeFileBegin);
 	size_t sizeToWrite = 0;
 	if (funcData.size() < oldFuncSize) {
 		nothingWasChanged = memcmp(funcPtr, funcData.data(), funcData.size()) == 0;
